# specs/openapi/MULO_OpenAPI_v2.7.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MULO API",
    "version": "2.7",
    "description": "MULO 통합 API 계약 기준 OpenAPI 문서입니다.\n\n- 현재 feature 구현과 기존 OpenAPI v2.6 계약을 함께 보존합니다.\n- 각 작업의 x-implementation-status는 implemented, contract-only, excluded-v1 중 하나입니다.\n- 홈 인기 자물쇠는 조회 시점 기준 최근 7일 데이터를 사용합니다.\n- 내 자물쇠 조회와 사용자 자물쇠 지역 그룹에는 7일 제한을 적용하지 않습니다.\n- 법정동 조회가 정상 완료됐지만 결과가 없으면 legalDongCode/legalDongName을 null로 허용합니다.\n- 커서 기반 목록 API size는 기본 20, 최소 1, 최대 100이며 채팅 메시지 내역만 기본 30입니다.\n- Refresh/CSRF Cookie의 SameSite 최종 정책은 배포 구조에 따라 확정하며 현재 구현은 Strict입니다.\n- 서버 URL은 아직 확정되지 않아 지정하지 않았습니다."
  },
...
  "paths": {
...
    "/api/recommendations/playlists": {
      "get": {
        "tags": [
          "RECOMMEND / 추천"
        ],
        "summary": "현재 추천 플레이리스트 조회",
        "operationId": "getApiRecommendationsPlaylists",
        "description": "사용자당 현재 추천 플레이리스트 1개. 아직 생성된 플레이리스트가 없으면 playlist=null.\n\n### 미정 정책\n- 필수\n없음\n\n선택\n없음\n\n정책 및 설명\n- 현재 로그인한 사용자의 현재 추천 플레이리스트를 조회한다.\n- 인증된 활성 사용자만 사용할 수 있다.\n- 사용자당 현재 추천 플레이리스트는 최대 1개만 유지한다.\n- 아직 생성된 추천 플레이리스트가 없으면 오류가 아니라 정상 상태로 처리하며 playlist=null을 반환한다.\n- 현재 플레이리스트가 존재하면 저장된 플레이리스트와 포함된 음악 목록을 반환한다.\n- 추천 생성 당시 사용한 날씨, 기온, 시간 정보는 추천 입력 컨텍스트로만 사용하며 playlist 리소스에 저장하지 않는다.\n- 따라서 조회 응답에도 weatherCondition, temperature, time을 포함하지 않는다.: 상태=현재 추천 플레이리스트 조회API는 Request Body, Path Parameter, Query Parameter가 없으므로 400은 필요없다",
        "responses": {
          "200": {
            "description": "리소스가 없는 상태도 현재 추천 상태 조회의 정상 결과이므로 200.\n\n추가 케이스: 현재 추천 플레이리스트가 존재하여 정상 반환하므로 200.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "playlist": {
                          "nullable": true
                        }
                      }
                    }
                  }
                },
                "example": {
                  "message": "현재 추천 플레이리스트 조회 성공",
                  "data": {
                    "playlist": null
                  }
                }
              }
            }
          },
          "401": {
            "description": "개인화 추천 조회이므로 인증이 필요하다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "로그인이 필요합니다."
                }
              }
            }
          },
          "500": {
            "description": "MULO 백엔드 내부의 예상하지 못한 처리 실패이므로 500.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "서버 내부 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                }
              }
            }
          }
        },
        "x-implementation-status": "implemented"
      },
      "post": {
        "tags": [
          "RECOMMEND / 추천"
        ],
        "summary": "새 추천 플레이리스트 생성",
        "operationId": "postApiRecommendationsPlaylists",
        "description": "필수\nlatitude: Double\nlongitude: Double\n\n동기 처리. 사용자의 기록 + 현재 위치 주변 장소의 자물쇠 + 현재 위치/장소 + 날씨 + 기온 + 시간을 추천 컨텍스트로 사용한다.\n\n정책 및 설명\n- 개인 기록과 주변 기록에는 최소 개수 제한이 없으며 0건이어도 추천할 수 있다.\n- 현재 위치의 날씨와 기온은 필수 컨텍스트다.\n- 동일 KST 날짜·격자의 weather_grid_forecasts 캐시를 우선 조회하고 당일 캐시가 없으면 기상청 단기예보 getVilageFcst를 조회한다.\n- 최초 적재는 목표 예보시각을 포함하는 가장 최근 후보 회차를 시도하고, 실패·미포함이면 바로 이전 후보 회차를 한 번만 시도한다. 모두 실패해 날씨·기온을 확보하지 못하면 502 WEATHER_API_ERROR로 추천을 생성하지 않는다.\n- 새 추천 생성에 성공한 경우에만 기존 현재 플레이리스트를 교체하며 실패하면 기존 플레이리스트를 유지한다.\n- 추천 생성 당시 날씨·기온·시간은 입력 컨텍스트로만 사용하고 playlist에 저장하거나 GET 응답으로 반환하지 않는다.\n- 같은 날짜·격자의 최초 성공 예보 묶음은 당일 유지하며, 더 최신 발표본으로 갱신하거나 외부 API를 재호출하지 않는다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "latitude",
                  "longitude"
                ],
                "properties": {
                  "latitude": {
                    "type": "number",
                    "format": "double",
                    "minimum": -90,
                    "maximum": 90,
                    "description": "현재 위치의 위도"
                  },
                  "longitude": {
                    "type": "number",
                    "format": "double",
                    "minimum": -180,
                    "maximum": 180,
                    "description": "현재 위치의 경도"
                  }
                }
              },
              "example": {
                "latitude": 37.5665,
                "longitude": 126.978
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "새 추천 플레이리스트 리소스가 생성되므로 201.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "playlist": {
                          "type": "object",
                          "properties": {
                            "recommendationPlaylistId": {
                              "type": "integer",
                              "format": "int64"
                            },
                            "tracks": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "musicTrackId": {
                                    "type": "integer",
                                    "format": "int64"
                                  },
                                  "title": {
                                    "type": "string"
                                  },
                                  "artistName": {
                                    "type": "string"
                                  },
                                  "externalUrl": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "message": "추천 플레이리스트 생성 성공",
                  "data": {
                    "playlist": {
                      "recommendationPlaylistId": 71,
                      "tracks": [
                        {
                          "musicTrackId": 3,
                          "title": "비 오는 날엔",
                          "artistName": "헤이즈",
                          "externalUrl": "..."
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "추천 컨텍스트 생성을 위한 현재 위치 좌표가 누락되어 요청 검증에 실패했으므로 400.\n\n추가 케이스: latitude 또는 longitude가 유효한 좌표 범위를 만족하지 않으므로 400.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "code": "INVALID_INPUT_VALUE",
                  "message": "입력값을 확인해주세요.",
                  "errors": [
                    {
                      "field": "latitude",
                      "code": "LATITUDE_REQUIRED",
                      "message": "현재 위도가 필요합니다."
                    },
                    {
                      "field": "longitude",
                      "code": "LONGITUDE_REQUIRED",
                      "message": "현재 경도가 필요합니다."
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "추천 플레이리스트 생성은 현재 사용자의 기록과 개인화 정보를 사용하는 기능이므로 유효한 Access Token이 없거나 인증에 실패한 경우 401.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "로그인이 필요합니다."
                }
              }
            }
          },
          "502": {
            "description": "현재 위치의 날씨·기온을 확보하지 못해 추천을 생성할 수 없으므로 502.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "WEATHER_API_ERROR",
                  "message": "날씨 정보를 불러오는 중 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                }
              }
            }
          },
          "500": {
            "description": "MULO 백엔드 내부의 예상하지 못한 처리 실패이므로 500.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "서버 내부 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                }
              }
            }
          }
        },
        "x-implementation-status": "implemented",
        "x-csrf-required": true
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access Token 인증이 필요한 API에서 사용"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "HttpOnly Refresh Token Cookie. Path=/api/auth, 운영 Secure=true. SameSite 최종 정책은 배포 구조에 따라 확정하며 현재 구현은 Strict."
      },
      "xsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-XSRF-TOKEN",
        "description": "GET /api/csrf가 발급한 XSRF-TOKEN Cookie의 원본 값을 상태 변경 요청에 전달한다. POST /api/users/signup은 예외다."
      }
    },
...
  }
}